+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### 1.0.2 - 2019-08-30
+
+- Added admin log for deleted users.
+- Added option for email message to deleted users.
+
+
 ### 1.0.1 - 2019-08-09
 
 - Added option to show message after register user.